# Optimize CONCAT(CONCAT(a, b), c)