Optimize CONCAT(CONCAT(a, b), c) reported in https://github.com/benhoyt/goawk/issues/91

Added a ConcatN opcode. Could have used MultiIndex opcode (I tried and it works
as is) but "concat" should not be using a separator.