@benhoyt 2022-02-03
Or maybe I'll get to it while the iron's hot. :-) Added a couple of review comments. Could you also please add the benchmark I mentioned and then post your before/after results of running all three benchmarks (ideally with `-count=5` and showing the diff using the benchstat tool). Thanks!