@raff 2022-02-04
Let me know if you want to go with a single Concat that supports 2+ parameters or keep 2 separate ones, given the benchmark results.