@benhoyt 2022-02-04
@raff Thanks for the updates. Yeah, I think having a separate ConcatN makes sense with those numbers. Maybe rename `Concat` to `Concat2`?

Could you also please post your before/after results for the other BenchmarkConcat benchmarks?