@raff 2022-02-04
Do you want Concat2 and ConcatN or Concat2 and Concat ?

I just figured out where to get benchstat, I'll post the before/after results in a bit.

Not sure of why when enabling/disabling the original concat I get a ~4% difference in ConcatSmall:

```
ConcatTwo-16             98.3ns ± 0%  84.0ns ± 6%  -14.58%  (p=0.008 n=5+5)
ConcatSmall-16            250ns ± 1%   259ns ± 2%   +3.91%  (p=0.008 n=5+5)
ConcatLarge-16            713µs ± 2%   712µs ± 5%     ~     (p=0.421 n=5+5)
```