@benhoyt 2022-02-04
Let's go with `Concat2` and `ConcatMulti`. Could you please rename `MultiIndex` to `IndexMulti` while you're at it, to be consistent with that ordering?