@raff 2022-02-04
This is the output of benchstat.sh. Not sure why there is a bunch of tests that every time I run it show up as slower than before (and completely unrelated to string concatenation):

```
name                     old time/op  new time/op   delta
ForInLoop-16             1.19µs ± 6%   1.33µs ± 2%  +11.44%  (p=0.008 n=5+5)
BuiltinGsub-16           12.6µs ± 0%   14.0µs ± 7%  +10.87%  (p=0.016 n=4+5)
BuiltinIndex-16           834ns ± 1%    904ns ± 3%   +8.31%  (p=0.008 n=5+5)
ForLoop-16               1.77µs ± 2%   1.92µs ± 7%   +8.24%  (p=0.016 n=5+5)
GlobalVars-16             200ns ± 6%    215ns ± 3%   +7.59%  (p=0.008 n=5+5)
BuiltinSplitRegex-16     68.6µs ± 0%   73.5µs ± 8%   +7.22%  (p=0.008 n=5+5)
LocalVars-16              221ns ± 5%    236ns ± 3%   +6.80%  (p=0.016 n=5+5)
BuiltinMatch-16          2.09µs ± 2%   2.21µs ± 4%   +5.82%  (p=0.016 n=5+5)
BinaryOperators-16        322ns ± 0%    341ns ± 1%   +5.74%  (p=0.008 n=5+5)
BuiltinSubAmpersand-16   10.5µs ± 1%   11.0µs ± 5%   +5.15%  (p=0.008 n=5+5)
BuiltinSubstr-16          568ns ± 1%    597ns ± 6%   +5.08%  (p=0.016 n=5+5)
IfStatement-16            107ns ± 1%    112ns ± 1%   +4.56%  (p=0.008 n=5+5)
CondExpr-16               106ns ± 2%    110ns ± 3%   +4.47%  (p=0.008 n=5+5)
BuiltinSub-16            10.6µs ± 1%   11.0µs ± 4%   +4.05%  (p=0.016 n=5+5)
ArrayOperations-16        434ns ± 1%    442ns ± 1%   +1.77%  (p=0.008 n=5+5)
Comparisons-16            245ns ± 0%    250ns ± 2%   +1.70%  (p=0.008 n=5+5)
AugAssign-16              755ns ± 1%    766ns ± 1%   +1.49%  (p=0.008 n=5+5)
IncrDecr-16               114ns ± 4%    110ns ± 7%     ~     (p=0.222 n=5+5)
SimpleBuiltins-16         595ns ± 6%    564ns ±11%     ~     (p=0.222 n=5+5)
BuiltinLength-16          334ns ± 2%    328ns ± 4%     ~     (p=0.278 n=5+5)
BuiltinSplitSpace-16     28.4µs ± 2%   28.7µs ± 6%     ~     (p=0.730 n=5+5)
BuiltinGsubAmpersand-16  12.8µs ± 2%   13.0µs ± 2%     ~     (p=0.095 n=5+5)
BuiltinSprintf-16        7.20µs ± 2%   7.23µs ± 4%     ~     (p=1.000 n=5+5)
RecursiveFunc-16         11.2µs ± 4%   11.3µs ± 6%     ~     (p=0.548 n=5+5)
FuncCall-16              1.99µs ± 1%   2.01µs ± 3%     ~     (p=0.690 n=5+5)
NativeFunc-16            8.08µs ± 2%   8.40µs ± 7%     ~     (p=0.056 n=5+5)
SimplePattern-16         97.8ns ± 3%  101.5ns ± 5%     ~     (p=0.151 n=5+5)
GetField-16               316ns ± 3%    319ns ± 4%     ~     (p=0.548 n=5+5)
SetField-16               640ns ± 0%    642ns ± 2%     ~     (p=0.730 n=4+5)
RegexMatch-16             856ns ± 0%    874ns ± 2%     ~     (p=0.111 n=4+5)
Assign-16                 202ns ± 2%    200ns ± 1%     ~     (p=0.341 n=5+5)
ConcatSmall-16            517ns ± 0%    263ns ± 4%  -49.15%  (p=0.008 n=5+5)
ConcatLarge-16           5.22ms ± 1%   0.71ms ± 1%  -86.46%  (p=0.008 n=5+5)
[Geo mean]               1.46µs        1.28µs       -12.52%
```