@benhoyt 2022-02-04
Yeah, benchmarks are funny sometimes. You have to run them on a quiet machine, and even then there can be fluctuation due to the compiler moving things around, or the x86 architecture being (too) clever. But here's what I'm seeing, and it's only split/sub that have slowed down a little, so I'm comfortable with this:

```
name                    old time/op  new time/op  delta
BuiltinSplitRegex-8     87.7µs ± 0%  94.7µs ± 9%   +7.98%  (p=0.008 n=5+5)
BuiltinSplitSpace-8     35.1µs ± 0%  37.6µs ± 0%   +7.10%  (p=0.008 n=5+5)
BuiltinGsubAmpersand-8  16.2µs ± 0%  17.3µs ± 6%   +6.56%  (p=0.008 n=5+5)
BuiltinSubAmpersand-8   13.6µs ± 0%  14.4µs ± 2%   +5.85%  (p=0.008 n=5+5)
BuiltinSub-8            13.6µs ± 0%  14.3µs ± 3%   +5.45%  (p=0.008 n=5+5)
BuiltinGsub-8           16.2µs ± 0%  16.6µs ± 0%   +2.38%  (p=0.016 n=5+4)
SetField-8               810ns ± 0%   829ns ± 1%   +2.33%  (p=0.008 n=5+5)
BuiltinMatch-8          2.71µs ± 1%  2.69µs ± 3%     ~     (p=0.635 n=5+5)
BuiltinSprintf-8        9.23µs ± 0%  9.27µs ± 1%     ~     (p=0.667 n=5+5)
RecursiveFunc-8         14.9µs ± 0%  14.7µs ± 1%   -1.21%  (p=0.008 n=5+5)
BinaryOperators-8        413ns ± 0%   407ns ± 1%   -1.67%  (p=0.008 n=5+5)
NativeFunc-8            10.8µs ± 0%  10.5µs ± 0%   -2.44%  (p=0.008 n=5+5)
FuncCall-8              2.76µs ± 2%  2.66µs ± 1%   -3.46%  (p=0.008 n=5+5)
CondExpr-8               151ns ± 0%   144ns ± 0%   -4.35%  (p=0.008 n=5+5)
GetField-8               435ns ± 2%   416ns ± 1%   -4.40%  (p=0.008 n=5+5)
IfStatement-8            152ns ± 0%   143ns ± 0%   -5.53%  (p=0.008 n=5+5)
SimpleBuiltins-8         746ns ± 0%   703ns ± 1%   -5.87%  (p=0.008 n=5+5)
RegexMatch-8            1.11µs ± 4%  1.05µs ± 3%   -5.89%  (p=0.032 n=5+5)
BuiltinIndex-8          1.11µs ± 1%  1.04µs ± 0%   -6.05%  (p=0.008 n=5+5)
AugAssign-8              982ns ± 0%   920ns ± 0%   -6.32%  (p=0.008 n=5+5)
ArrayOperations-8        565ns ± 0%   528ns ± 1%   -6.67%  (p=0.008 n=5+5)
SimplePattern-8          134ns ± 0%   125ns ± 1%   -6.69%  (p=0.008 n=5+5)
BuiltinLength-8          429ns ± 0%   400ns ± 2%   -6.70%  (p=0.008 n=5+5)
ForLoop-8               2.50µs ± 0%  2.33µs ± 3%   -6.78%  (p=0.008 n=5+5)
LocalVars-8              300ns ± 2%   279ns ± 1%   -6.93%  (p=0.008 n=5+5)
Comparisons-8            321ns ± 0%   298ns ± 0%   -7.27%  (p=0.029 n=4+4)
GlobalVars-8             269ns ± 1%   249ns ± 3%   -7.58%  (p=0.008 n=5+5)
Assign-8                 261ns ± 0%   240ns ± 1%   -7.76%  (p=0.008 n=5+5)
BuiltinSubstr-8          765ns ± 0%   705ns ± 1%   -7.92%  (p=0.008 n=5+5)
ForInLoop-8             1.78µs ± 4%  1.63µs ± 3%   -8.08%  (p=0.008 n=5+5)
IncrDecr-8               148ns ± 0%   136ns ± 1%   -8.22%  (p=0.016 n=4+5)
ConcatSmall-8            691ns ± 1%   341ns ± 3%  -50.72%  (p=0.008 n=5+5)
ConcatLarge-8           8.35ms ± 2%  1.10ms ± 3%  -86.88%  (p=0.008 n=5+5)
[Geo mean]              1.94µs       1.60µs       -17.75%
```