@benhoyt 2022-02-04
Thanks again for the contribution, much appreciated! I'll merge this shortly. Actually, looking at this diff now I think I'll revert `Concat2` back to just `Concat` like all the other binary ops (I'll do this after the merge, so don't worry).