@benhoyt 2022-02-04
Final tweaks and updating benchmark results (from the run on my machine): https://github.com/benhoyt/goawk/commit/bd053c958b3dd1a4460768d1b344ee610059edd6